requirements.txt
click==8.0.3
datasets==1.17.0
hydra-core==1.1.1
numpy==1.22.0
pandas==1.3.5
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
setuptools==60.2.0
torch==1.8.1
tqdm==4.62.3
transformers==4.15.0
